Decouple content links service from events and dealers
